Refactor to DRY
